AgregadasRefs
Acomodadas casi todas las refs.
Faltan:
Some authors \cite{buscar} have proposed to add a random perturbations
to avoid this drawback of the Tent map.

The first step to quantify the statistical properties of the values
(amplitude statistics) of a time series $\{x_i,~(i=1,...,N)\}$ using
information theory is to determine the concomitant PDF.
This is an issue studied in detail in previous papers \cite{aka varios}.

The Perron-Frobenius operator and its corresponding eigenvalues and
eigenfunctions may be also be analytically obtained for this map
\cite{tent}.

Falta acomodar las abcisas de los plots